+++ iface/dynamic.go
@@ -14,7 +14,7 @@ import (
 type IDynamic interface {
 	Quit()
 	GetConf() *gvar.GroupConf
-	GetGroup(groupId int32) (IGroup, error)
-	Cast(groupId int32, msg *gvar.MsgData) error
+	GetGroup(groupId int64) (IGroup, error)
+	Cast(groupId int64, msg *gvar.MsgData) error
 	Entry(conn *websocket.Conn)
 }
